react to feedback
