+++ docs/core/deploying/trimming/snippets/ConsoleApp1/Program.cs
@@ -1,3 +1,4 @@
 ﻿using System;
+using MyLibrary1;
 
 Console.WriteLine(MyLib.MyClass.getMax(1, 2));
